Раздробил код на модули
