+++ js/generate-data-stub/generate-post-stub/index.js
@@ -1,0 +1,5 @@
+import { generatePostsListStub } from './generate-post';
+
+export {
+  generatePostsListStub
+};
