+++ js/generate-data-stub/index.js
@@ -1,0 +1,9 @@
+import { generatePostsListStub } from './generate-post-stub';
+
+const generatePosts = {
+  generatePostsListStub
+};
+
+export {
+  generatePosts
+};
