+++ js/utills/index.js
@@ -1,0 +1,21 @@
+import {
+  getRandomIntegerInInterval,
+  getRandomItemInArray
+} from './tools/random';
+import {
+  createUniqueNumberInIntervalGenerator
+} from './tools/number-generator';
+
+const randomTools = {
+  getRandomIntegerInInterval,
+  getRandomItemInArray
+};
+
+const numberGeneratorTools = {
+  createUniqueNumberInIntervalGenerator
+};
+
+export {
+  randomTools,
+  numberGeneratorTools
+};
